migrated from http to gRPC for client requests
