+++ src/client.rs
@@ -1,6 +1,6 @@
 use core::task;
 use hello_world::seafoam_client::SeafoamClient;
-use hello_world::GetRequest;
+use hello_world::{GetRequest, SetRequest};
 use std::borrow::Borrow;
 use std::sync::Arc;
 use std::time::*;
@@ -25,18 +25,24 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let permit = semaphore.clone().acquire_owned().await.unwrap();
         tokio::spawn(async move {
             loop {
-                if let Ok(channel) = channel::Endpoint::from_static("http://[::1]:50051")
+                if let Ok(channel) = channel::Endpoint::from_static("http://[::1]:3000")
                     .connect()
                     .await
                 {
                     let _p = permit;
                     let mut client = SeafoamClient::new(channel);
-                    for i in 0..25 {
+                    for i in 0..24 {
                         let request = tonic::Request::new(GetRequest {
                             key: "hello".to_string(),
                         });
                         let response = client.get(request).await.unwrap();
                     }
+                    let request = tonic::Request::new(SetRequest {
+                        key: "hello".to_string(),
+                        value: "{\"name\": \"world\"}".to_string(),
+                    });
+                    let response = client.set(request).await.unwrap();
+                    //println!("{:?}", response.into_inner());
                     return;
                 }
             }
