+++ src/main.rs
@@ -8,7 +8,6 @@ use std::time::Instant;
 use tonic::transport::Server;
 
 use flashmap::{self, ReadHandle};
-use grpc::seafoam_client::SeafoamClient;
 use hyper::service::service_fn;
 use hyper::Request;
 use hyper_util::rt::TokioIo;
@@ -17,92 +16,10 @@ use tokio::net::TcpListener;
 use tokio::sync::{watch, RwLock};
 use tonic::Response;
 
-use seafoam::{
-    kv_store,
-    raft::{log_manager, raft_state_manager, Data, RaftCore, RaftState},
-};
+use seafoam::raft::{log_manager, raft_state_manager, Data, RaftCore, RaftState};
 
-pub mod grpc {
-    tonic::include_proto!("seafoam");
-}
-
-use grpc::seafoam_server::{Seafoam, SeafoamServer};
-use grpc::*;
-
-pub struct MyServer {
-    db: ReadHandle<String, Data>,
-    log: Arc<RwLock<Vec<LogEntry>>>,
-    internal_state: Arc<RwLock<RaftCore>>,
-    state_receiver: watch::Receiver<RaftState>,
-}
-
-#[tonic::async_trait]
-impl Seafoam for MyServer {
-    async fn get(
-        &self,
-        request: tonic::Request<GetRequest>,
-    ) -> Result<Response<GetReply>, tonic::Status> {
-        let key: String = request.into_inner().key;
-        let start = Instant::now();
-        let data = self.db.guard().get(&key).cloned();
-        let end = start.elapsed().as_millis();
-        if end > 10 {
-            println!("{}", end);
-        }
-
-        match data {
-            None => Ok(Response::new(GetReply { value: None })),
-            data => {
-                let resp = GetReply {
-                    value: Some(serde_json::to_string(&data).unwrap()),
-                };
-                Ok(Response::new(resp))
-            }
-        }
-    }
-
-    async fn set(
-        &self,
-        request: tonic::Request<SetRequest>,
-    ) -> Result<Response<SetReply>, tonic::Status> {
-        // Make sure to implement proper logic to forward the request to the leader of the Raft cluster
-        let state = self.state_receiver.borrow().clone();
-        match state {
-            RaftState::Leader(term) => {
-                let message = request.into_inner();
-                let key = message.key;
-                let value = message.value;
-                let mut write_lock = self.log.write().await;
-
-                let index = write_lock.len() + 1;
-                let log_entry = LogEntry::Insert {
-                    key: key,
-                    data: serde_json::from_str(&value).unwrap(),
-                    index: index,
-                    term: term,
-                };
-
-                write_lock.push(log_entry);
-                drop(write_lock);
-
-                // Figure out an optimal way to trigger a response
-                // while self.internal_state.read().await.max_committed < index {
-                //     tokio::time::sleep(tokio::time::Duration::from_millis(10)).await;
-                // }
-
-                Ok(Response::new(SetReply {}))
-            }
-            RaftState::Follower(_, Some(leader), _) => {
-                let leader = leader.clone();
-                let mut client = SeafoamClient::connect(format!("http://{}", leader))
-                    .await
-                    .unwrap();
-                client.set(request).await
-            }
-            _ => Err(tonic::Status::unavailable("No leader currently elected")),
-        }
-    }
-}
+use seafoam::grpc::seafoam_server::SeafoamServer;
+use seafoam::MyServer;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
@@ -146,26 +63,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     drop(writer);
 
     let (writer, reader) = flashmap::new();
-    // This is wyatt's bad idea
-    let greeter = MyServer {
-        db: reader.clone(),
-        log: log.clone(),
-        internal_state: internal_state.clone(),
-        state_receiver: state_receiver.clone(),
-    };
-
-    tokio::spawn(async {
-        let add = "[::1]:50051".parse().unwrap();
-        Server::builder()
-            .add_service(SeafoamServer::new(greeter))
-            .serve(add)
-            .await
-    })
-    .await?;
-    // This is normal code
-
-    let listener = TcpListener::bind(addr).await?;
-    println!("Listening on http://{}", addr);
 
     let l = log.clone();
     let i = internal_state.clone();
@@ -179,22 +76,23 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         args[2].parse::<u16>().unwrap(),
     ));
 
-    loop {
-        let (stream, _) = listener.accept().await?;
-        let io = TokioIo::new(stream);
+    // This is wyatt's bad idea
+    let greeter = MyServer {
+        db: reader.clone(),
+        log: log.clone(),
+        internal_state: internal_state.clone(),
+        state_receiver: state_receiver.clone(),
+    };
 
-        let new_reader = reader.clone();
-        let v = log.clone();
-        let s = state_receiver.clone();
-        let service = service_fn(move |req: Request<hyper::body::Incoming>| {
-            kv_store(req, new_reader.clone(), v.clone(), s.clone())
-        });
-        let http = http1::Builder::new();
-
-        tokio::task::spawn(async move {
-            if let Err(err) = http.serve_connection(io, service).await {
-                println!("Error serving connection: {:?}", err);
-            }
-        });
-    }
+    println!("Listening on http://{}", addr);
+    Ok(tokio::spawn(async move {
+        let addr = format!("[::1]:{}", client_port);
+        let add = addr.parse().unwrap();
+        Server::builder()
+            .add_service(SeafoamServer::new(greeter))
+            .serve(add)
+            .await
+    })
+    .await??)
+    // This is normal code
 }
